+++ jira.go
@@ -15,7 +15,7 @@ type issue struct {
 	Status           string   `json:"status"`
 	Assignee         string   `json:"assignee"`
 	AssigneeImageURL string   `json:"assigneeImageURL"`
-	Estimate         int      `json:"estimate"` // note that this doesn't differentiate between '0' and unset
+	Estimate         float64  `json:"estimate"` // note that this doesn't differentiate between '0' and unset
 	Priority         string   `json:"priority"`
 	PriorityImageURL string   `json:"priorityImageURL"`
 	Labels           []string `json:"labels"`
